add Main component
